+++ src/main/java/com/playus/userservice/domain/user/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.playus.userservice.domain.user.controller;
 
-import com.playus.userservice.domain.user.dto.UserWithdrawResponse;
+import com.playus.userservice.domain.user.dto.withdraw.UserWithdrawResponse;
 import com.playus.userservice.domain.user.dto.nickname.NicknameRequest;
 import com.playus.userservice.domain.user.dto.nickname.NicknameResponse;
 import com.playus.userservice.domain.user.specification.UserControllerSpecification;
