+++ src/main/java/com/playus/userservice/domain/user/dto/profile/UserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.playus.userservice.domain.user.dto;
+package com.playus.userservice.domain.user.dto.profile;
 
 public record UserInfoResponse(
         String nickname,
