+++ src/main/java/com/playus/userservice/domain/user/dto/withdraw/UserWithdrawResponse.java
@@ -1,3 +1,3 @@
-package com.playus.userservice.domain.user.dto;
+package com.playus.userservice.domain.user.dto.withdraw;
 
 public record UserWithdrawResponse(boolean success, String message) {}
