+++ src/main/java/com/playus/userservice/domain/user/service/UserTagReadService.java
@@ -1,6 +1,6 @@
 package com.playus.userservice.domain.user.service;
 
-import com.playus.userservice.domain.user.dto.UserTagSummaryResponse;
+import com.playus.userservice.domain.user.dto.review.UserTagSummaryResponse;
 import com.playus.userservice.domain.user.document.UserTagDocument;
 import com.playus.userservice.domain.user.document.TagDocument;
 import com.playus.userservice.domain.user.repository.read.TagDocumentRepository;
